fix(ci): unblock main eresolve and full hygiene sweep
