+++ package.json
@@ -41,7 +41,7 @@
     "discord.js": "^14.27.0",
     "protobufjs": "8.8.0",
     "telegraf": "^4.16.3",
-    "zod": "^4.5.4"
+    "zod": "^3.22.4"
   },
   "devDependencies": {
     "@cloudflare/vitest-pool-workers": "0.22.0",
@@ -49,7 +49,7 @@
     "@playwright/test": "^1.63.0",
     "@types/js-yaml": "^4.0.9",
     "@types/node": "^26.4.1",
-    "@vitest/coverage-v8": "^5.0.0",
+    "@vitest/coverage-v8": "^4.1.11",
     "artillery": "^2.0.34",
     "husky": "^9.1.7",
     "js-yaml": "^5.4.1",
@@ -57,7 +57,7 @@
     "miniflare": "4.20260730.0",
     "prettier": "^3.9.6",
     "typescript": "^7.0.2",
-    "vitest": "^5.0.0",
+    "vitest": "^4.1.11",
     "wrangler": "^4.129.0"
   },
   "overrides": {
